requirements.txt
Django==4.1.7
django_rest_swagger==2.2.0
djangorestframework==3.14.0
celery~=5.2.7